requirements.txt
Flask>=2.0
requests>=2.25
beautifulsoup4>=4.9
ics>=0.7
python-dateutil>=2.8
APScheduler>=3.6
pytz>=2020.1
python-dotenv>=0.19
